backend/requirements.txt
fastapi
uvicorn[standard]
mediapipe
opencv-python
openai
supabase
python-multipart
numpy
python-dotenv
pydantic
lunar-python
pyjwt
